easy_buy_server/requirements.txt
Django~=4.2.13
dj-rest-auth
drf-yasg
django-cors-headers
djangorestframework
django-allauth
